finally get rid of old gomidi/midi version. the price we pay for the moment: - smfimport is disabled, but it did never work well - playing is not interruptable at the moment. this needs to be fixed in gomidi/midi

+++ items/event.go
@@ -212,7 +212,7 @@ func (e *Event) Dup() *Event {
 }
 
 // AbsPosTicks returns the absolute Position in ticks
-func (e *Event) AbsPosTicks(ticks4th uint32, trackDelay int) uint {
+func (e *Event) AbsPosTicks(ticks4th uint16, trackDelay int) uint {
 	pos := uint(ticks4th) / 8 * e.Position
 	if e.PosShift == 0 {
 		res := int(pos) + trackDelay
